requirements.txt
django-cacheops==7.2
pillow==12.2.0
